standardize semantic kernel agent names
